Move HttpMetrics into the weco.http.monitoring namespace
